Move validate into another package - vallie
Remove logging
Update tests
